samples/03-github-operations/requirements.txt
toolweaver>=0.13.7
requests>=2.32.0
python-dotenv>=1.0.0
